[signon] Continue reading from plugins after handling response if necessary
There may be more than one response from the plugin; for example, during
a refresh, you receive a store and result response. If both arrive
simultaniously, readyRead() will not be signalled again after handling
the first response. The request will then be deadlocked waiting for a
response that is sitting in the file buffer.

The immediate solution is simply to read again if there's more to be
read after handling a response.